updated tilt calibration as well as rounding

+++ modulus.c
@@ -3,5 +3,5 @@
 # include "modulus.h"
 
 int modulo(int dividend, int divisor) {
-	return (int) (dividend - divisor * (int) floor(dividend / (float) divisor));
+	return dividend - divisor * ((int) floor(dividend / (float) divisor));
 }
